+++ CHANGES.rst
@@ -5,6 +5,9 @@ Version 2.1.2
 
 Unreleased
 
+-   The development server does not set ``Transfer-Encoding: chunked``
+    for 1xx, 204, 304, and HEAD responses. :issue:`2375`
+
 
 Version 2.1.1
 -------------